+++ next.config.js
@@ -5,23 +5,23 @@ const nextConfig = {
     serverActions: true,
   },
   output: "standalone",
+  staticPageGenerationTimeout: 180, // tingkatkan menjadi 3 menit
   // eslint: {
   //   ignoreDuringBuilds: true,
   // },
-  async headers() {
-    return [
-      {
-        source: "/(.*)",
-        headers: [
-          {
-            key: "Cache-Control",
-            value: "no-store, max-age=0",
-          },
-        ],
-      },
-    ];
-  },
-  
+  // async headers() {
+  //   return [
+  //     {
+  //       source: "/(.*)",
+  //       headers: [
+  //         {
+  //           key: "Cache-Control",
+  //           value: "no-store, max-age=0",
+  //         },
+  //       ],
+  //     },
+  //   ];
+  // },
 };
 
 module.exports = nextConfig;