+++ src/app/dev/admin/forum/sub-detail/report-posting/[id]/page.tsx
@@ -3,19 +3,19 @@ import { adminForum_getListReportPostingById } from "@/app_modules/admin/forum/f
 import { adminForum_getOnePostingById } from "@/app_modules/admin/forum/fun/get/get_one_posting_by_id";
 
 export default async function Page({ params }: { params: { id: string } }) {
-  let postingId = params.id;
-  const listReport = await adminForum_getListReportPostingById({
-    postingId: postingId,
-    page: 1,
-  });
+  // let postingId = params.id;
+  // const listReport = await adminForum_getListReportPostingById({
+  //   postingId: postingId,
+  //   page: 1,
+  // });
 
-  const dataPosting = await adminForum_getOnePostingById(postingId);
+  // const dataPosting = await adminForum_getOnePostingById(postingId);
 
   return (
     <>
       <AdminForum_HasilReportPosting
-        dataPosting={dataPosting as any}
-        listReport={listReport}
+        // dataPosting={dataPosting as any}
+        // listReport={listReport}
       />
     </>
   );